Rename nomis case notes to prison case notes (to match project name)
