+++ projects/prison-case-notes-to-probation/deploy/values-dev.yml
@@ -1,5 +1,5 @@
 environment_name: delius-test
 
 env:
-  INTEGRATIONS_CASE-NOTES_URL: https://offender-case-notes-dev.hmpps.service.justice.gov.uk
+  INTEGRATIONS_PRISON-CASE-NOTES_URL: https://offender-case-notes-dev.hmpps.service.justice.gov.uk
   SPRING_SECURITY_OAUTH2_CLIENT_PROVIDER_HMPPS-AUTH_TOKEN-URI: https://sign-in-dev.hmpps.service.justice.gov.uk/auth/oauth/token
