+++ projects/prison-case-notes-to-probation/deploy/values.yml
@@ -8,9 +8,9 @@ memory: 1024 # = 1 GB
 env: []
 
 secrets:
-  APPLICATIONINSIGHTS_CONNECTION_STRING: /application-insights/connection-string
-  SPRING_DATASOURCE_URL:                 /delius-database/jdbc-url
-  SPRING_DATASOURCE_USERNAME:            /prison-case-notes-to-probation/db-username
-  SPRING_DATASOURCE_PASSWORD:            /prison-case-notes-to-probation/db-password
-  INTEGRATIONS_CASE-NOTES_CLIENT-ID:     /prison-case-notes-to-probation/client-id
-  INTEGRATIONS_CASE-NOTES_CLIENT-SECRET: /prison-case-notes-to-probation/client-secret
+  APPLICATIONINSIGHTS_CONNECTION_STRING:        /application-insights/connection-string
+  SPRING_DATASOURCE_URL:                        /delius-database/jdbc-url
+  SPRING_DATASOURCE_USERNAME:                   /prison-case-notes-to-probation/db-username
+  SPRING_DATASOURCE_PASSWORD:                   /prison-case-notes-to-probation/db-password
+  INTEGRATIONS_PRISON-CASE-NOTES_CLIENT-ID:     /prison-case-notes-to-probation/client-id
+  INTEGRATIONS_PRISON-CASE-NOTES_CLIENT-SECRET: /prison-case-notes-to-probation/client-secret
