+++ projects/prison-case-notes-to-probation/src/main/kotlin/uk/gov/justice/digital/hmpps/config/FeignConfig.kt
@@ -15,18 +15,18 @@ import org.springframework.security.oauth2.client.OAuth2AuthorizeRequest
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClientManager
 import org.springframework.security.oauth2.core.OAuth2AccessToken
 import org.springframework.web.server.ResponseStatusException
-import uk.gov.justice.digital.hmpps.integrations.casenotes.CaseNotesClient
+import uk.gov.justice.digital.hmpps.integrations.prison.PrisonCaseNotesClient
 import uk.gov.justice.digital.hmpps.listener.ServicePrincipal
 import java.nio.charset.StandardCharsets
 
 @Configuration
-@EnableFeignClients(clients = [CaseNotesClient::class])
+@EnableFeignClients(clients = [PrisonCaseNotesClient::class])
 class FeignConfig(
     private val authorizedClientManager: OAuth2AuthorizedClientManager,
     private val servicePrincipal: ServicePrincipal
 ) {
     companion object {
-        const val REGISTRATION_ID = "case-notes"
+        const val REGISTRATION_ID = "prison-case-notes"
     }
 
     @Bean
